Generate interfaces, containers and builders
This class adds a code generator that generates interfaces, containers
and builders for the model types. This means, for example, that for the
"Vm" type the following classes will be created:

  V4Vm - The type interface.
  V4VmContainer - An implementation of the interface.
  V4VmBuilder - A builder for the type.

The interface generated for each type only contains methods for reading,
on order to have them as inmutable as possible. The "V4Vm" interface,
for example looks like this:

  public interface V4Vm {
    String getName();
    boolean hasName();
    ...
  }

The container class generated for each type supports reading and
writing, but it is intended for internal use by the builder classes
only. They look like this:

  public class V4VmContainer implements V4Vm {
    private String name;
    ...

    public String getName() {
      return name;
    }

    public void setName(String newName) {
      name = newName;
    }
    ...
  }

The builder classes are intended to build instances, using a method
chaining approach. The builder for the "Vm" type looks like this:

  public class V4VmBuilder {
    private String name;
    ...

    public V4VmBuilder setName(String newName) {
      name = newName;
      return this;
    }

    ...
    public V4Vm build() {
      V4VmContainer container = new V4VmContainer();
      container.setName(name);
      ...
      return container;
    }
  }

The new generator also creates a "V4Builders" class that contains static
methods intended for static import, and to avoid importing and
explicitly calling the constructors of the containers. For example:

  public class V4Builders {
    public V4Vm newVm() {
      return new V4VmBuilder();
    }
    ...
  }

The intended use of all these classes is something like this:

  V4Vm vm = newVm()
    .setId("123")
    .setName("myvm")
    .setMemory(1000)
    .setCpu(newCpu().setSpeed(1.0))
    .build();

These classes are generated but not used at the moment (except by a
little test). Later patches will start to change services to use them
instead of the classes generated by the XJC compiler from the XML
schema. Eventually the XJC generated code will be completely removed.

Change-Id: I45eee6e6fa15d758630055f1f8ffe5102ec04ffb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>