Add automation scripts
This patch adds the automation scripts.

The "build-artifacts.sh" scripts creates the tarball, builds the Maven
artifacts and generates the JSON and XML descriptions of the model. All
these artifacts are copied to the exported artifacts directory.

The "check-patch.sh" script runs the unit tests and findbugs.

The "check-merged.sh" script does nothing yet.

Change-Id: Iab06efc6a717bad3212113385eca441343d6a203
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>