Add tests module
This patch adds a new "tests" module that will contain the tests of the
metamodel.

The "src/main/java" directory of this new module will contain a dummy
model, to be used as the victim of the tests. Initially it will contain
only a few interfaces, like "Vm", "Cpu" and "Disk", but only with the
attributes required to have at least one of each type. In the future
this may be expanded, as required for new tests.

The "src/test/java" will contain the actual tests.

Change-Id: Ic355c2e720dcd6289b4795a5d2394c375c22d75d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>