Add option to enable generation of JAX-RS code
Currently the JAX-RS code is generated when the "--java" option is
specified. This forces the user of the tool to always generate the
JAX-RS interfaces and the rest of the Java code: all or nothing. To
avoid that this patch introduces a new "--jaxrs" option to enable
the generation of the JAX-RS code.

Change-Id: Icedc6ca88f6b15fcf2e58ca9b4b3b757fcfe5732
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>