Update version of commons-cli to 1.3.1
The version that we currently use, 1.3, doesn't handle correctly long
options that start with the same prefix, it considers them ambiguous.
For example, if both "--jaxrs" and "--jaxrs-package" are given it will
generate the following error message:

  Ambiguous option: '--jaxrs'  (could be: 'jaxrs', 'jaxrs-package')

This is a bug in commons-cli, which was fixed in 1.3.1:

  LongOpt falsely detected as ambiguous
  https://issues.apache.org/jira/browse/CLI-252

This patch updates the version of commons-cli to 1.3.1, to avoid this
issue.

Change-Id: I0820cc47e7e539d41ca24ee853b5a844fa60cd5a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>