Fix the calculation of MIME type in server
The model servlet calculates the response MIME type from extension of
the request URL. For example, if the request url ends in .json it should
use JSON, and if it is .xml it should use XML. But this calculation is
implemented incorrectly, so that it uses JSON for both .json and .xml.
This patch fixes that error.

Change-Id: I3f2d5e656a8aba39dcbf840e92f195e6945346f5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>