restapi: Use Enums In API entities
This patch changes the generation process of the API business
entities such that, from now on entities which in the model have
enum fields, will be generated with references to the enums, and
not to strings, as before.

Also, enums are now manually generated and JaxB annotations are
added to them by the metamodel infrastructure. Enums are still
written into the schema file, api.xsd, and then generated by
JaxB, but they are then overwritten by the metamodel, which
generates them differently.

Change-Id: I2a8ae72ef0b00a67a0dec01069145e7fd27311cc
Signed-off-by: Ori Liel <oliel@redhat.com>