Add built-in types to the model
Currently the built-in types, like "Fault" and "Action" are explicitly
added to the generated XML schema. But these built-in types are also
needed by other code generators, like the generators for the SDKs. This
patch moves this code to a utility class that adds the definition of
these types to the model, so it can be used by the generator of the XML
schema and also by the generators of the SDKs.

Change-Id: I29ef877f7cfa616a3610de21c1ed7d03fcd15e5c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>