Fix typo in the name of the OpenStack image providers locator
The name of this locator is currently "openstackImageProviers", there is
a missing "d". As a result the URL of the service in the engine is
"openstackimageproviers", which is wrong. This patch fixes the typo. The
engine will need to be fixed to use a version of the model that includes
this fix.

Change-Id: Id884e1d22923ec5c098cbbcd7487cdcc18cf138f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>