Remove VmService#Move
oVirt 3.0 had an API to move a VM from one storage domain to another.
oVirt 3.1 introduced the notion of a VM having multiple disks on
multiple storage domains, with the ability to move them independently,
rendering this API command meaningless, and it was indeed marked as
deprecated and preserved only in order to not break backwards
compatibility.

With the introduction of the 4.0 API, it can now safely be removed.

Change-Id: I835cd2c75097607268ebc84c285dd038c0386ca3
Signed-off-by: Allon Mureinik <amureini@redhat.com>