Add an enum type to the tests
There is currently an issue in the generation of the JSON support code
that isn't currently detected because the affected code is never
generated, but it will be generated in the future. This patch adds to
the test model an enum type, so that this will be detected if it happens
again in the future.

Change-Id: I0cc4723b28683bae94a067f24b786dc497be1f5b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>