Make main class generic
Currently the different tools available in the metamodel have each its
own "main" class that creates the CDI container and runs the tool. These
"main" classes are almost identical, the only difference is the name of
the tool class that they use. This makes original "Main" class generic,
so that it receives the name of the tool as a parameter, this way the
other copies can be removed.

Change-Id: I0333e6f3772a297632ed77f4b7f54d65d3d8ea17
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>