Don't generate full description when there is only summary
Currently the documentation generator generates the full description of
attributes, links, enum values and parameters even if they don't have
any documentation, or if they only have a summary. As this summary
(maybe empty) is already part of the summary of the struct, enum or
method, it isn't worth repeating it. This patch changes the
documentation generator so that it doesn't generate this repeated text.

Change-Id: I71fc2244770a290ac9a4ef0b8ce20fa1ec68dd21
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>