Added "reportStatus" to Storage Get operation
The "reportStatus" parameter indicates if the status of the LUNs
in the storage should be checked.
Checking the status of the LUN is a heavy weight operation and
this data is not always needed by the user.

This parameter will give the option to skip the status check
of the LUNs.

The default is 'true' for backward compatibility.

Change-Id: I396b5d937c45f275d53f5fa631f6f38f1c0f887f
Bug-Url: https://bugzilla.redhat.com/1308350
Signed-off-by: Fred Rolland <frolland@redhat.com>