Add annotations support
This patch adds support for annotations in the definitions of concepts.
These annotations are intended to add to the entities descriptions that
don't fit into the model. For example, we want to add to the model
information about what is the area/team responsible for a particular
service, to do so we will take advantage of annotations, like this:

  @Service
  @Area("Storage")
  public class DisksService {
    ...
  }

Tools will then be able to get this annotations from the model, and use
them for their own purposes.

Change-Id: Iec4e62394ea3fe4415b99a164beeed813ca913dd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>