Assign areas to services
This patch adds a new "Area" annotation that is intended to indicate
what area of the engine is a service related to. Initially these areas
are taken from the descriptions of the teams in bugzilla:

  https://bugzilla.redhat.com/page.cgi?id=fields.html#cf_ovirt_team

The assignment is made adding the "Area" annotation to all the service
specifications. For example, for the "Vms" service:

  @Area("Storage")
  public interface DisksService {
    ...
  }

The value of this annotation is used only for information and reporting
purposes, and it doesn't have any effect on the generated code or
documentation.

Change-Id: I2cb87cac5197afbbc9693a72ce3a5e1d32e52b44
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>