Generate a separate file for CSS
Currently the HTML documentation is generated with an inline CSS
stylesheet. But the default configuration of the Jenkins web server uses
the "Content Security Mechanism" to specify the stylesheets can only be
loaded from the same web server, using an HTTP header like this:

  X-Content-Security-Policy:
  "sandbox; default-src 'none'; img-src 'self'; style-src 'self';"

One of the effects of this header in most modern browsers is that the
inline stylesheet is ignored. The net result is that when the generated
HTML is downloaded from Jenkins directly the stylesheet isn't applied
and the documentation doesn't look as expected.

To avoid that issue this patch adds the "linkcss" attribute to the
generated AsciiDoc file, so that AsciiDoctor will generate the
stylesheet in a separate "asciidoctor.css" file, which is allowed by the
content security policy.

Change-Id: I81ee9cc36c836f0f05dc83b65612b026ecec46c8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>