Replace generic "Status" type with specific ones
Currently multiple types have an "status" attribute, and it is define as
a pair of "state" and "detail" strings containing in an "Status" struct.
This means that enums can't be used. To improve that this patch removes
the "Status" struct and replaces its usage with the corresponding enum
types, and an additional "*Detail" text attribute when it is really
needed, currently only for hosts and VMs.

Change-Id: I2f21851d132a5e8481c6c78cd239ff3bdbfc311f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>