Restore the "bootable" and "interface" disk attributes
This patch re-adds the disk "bootable" and "interface" attributes. These
attributes are intended to be removed as the change that introduces the
concept of "disk attachments", which replaces the concept of "disks" for
virtual machines. But removing these two attributes requries a lot of
change that interfere with other changes required in the model and in
the engine. To simplify those other changes this patch re-introduces the
attributes. They will be removed later, once the other model changes are
performed.

Change-Id: I21f332f85214ab87dc8e750b3c613f21317f0201
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>