remove MacPool from DataCenter
In 4.0 we move away from DataCenter concept, and to be able to do that, 
we need to move configuration previously set on DataCenter to finer 
level. Because of that from now on, mac pool is set on cluster level.
This is the coarsest setting, in future it's probable it will be able to 
set it on finer level as well.

Change-Id: I67e589f76e37b6e2f0fa12475c3899543d05fa2f
Signed-off-by: Martin Mucha <mmucha@redhat.com>