Use dash as section id separator
Currently the generated AsciiDoc file uses slashes to separate the
components of the identifiers of sections. For example, the identifier
for the section corresponding to the VmService#start method is the
following:

  services/vm/methods/start

This complicates processing of the AsciiDoc text for some tools. To
avoid that this patch changes the tool so that it uses dashes instead of
slashes, so the same identifier will look like this:

  services-vm-methods-start

Change-Id: I252c19a8a979ee6bda8914d1f254ade6bff16df0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>