Convert Javadoc tags into model annotations
This patch modifies the model analyzer so that it converts Javadoc tags
into model annotations. For example, if the Javadoc comment for a
concept contains the following text:

  /**
   * @author Jane Doe
   * @author John Doe
   */

The model will be populated with an annotation named "Author" with a
"value" parameter containing two values, "Jane Doe" and "John Doe".

Change-Id: I7138b575170e9325fa36c02ba32dabbdfc64a2e5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>