Fix several section references
The format of a section reference is the following:

  <<section_id,text>>

But it some places we reversed the section id and the text, which
results in links that don't look nice.

This patch fixes those incorrect section references.

Change-Id: I3ac33e5768d3bd351caeec0f3ae2784015997805
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 41b8c9f586f63a47ef23e23eaad424005ac87d74)