Change name of output parameter of AffinityGroupsService list interface
This patch changes name of the output parameter of AffinityGroupsService
list interface. It was named list() which is not consintent name in
terms of names of other services. We are changing the name to
groups() to be consistent.

Change-Id: I3fae31d41c8e9971e69d6d2081a66fb816cee7e1
Signed-off-by: Ondra Machacek <omachace@redhat.com>
(cherry picked from commit e91181bcbada6a2d2dd4ce648b6f57729f2ab251)