Add product and engine name attributes
This patch adds two AsciiDoctor attributes intended to contain the name
of the product and the name of the engine. The default values will be
the following:

  :product-name: oVirt
  :engine-name: oVirt Engine

These are intended for customization for downstream packages. For
example, Red Hat may wish to change these properties to the following
values:

  :product-name: Red Hat Virtualization
  :engine-name: Red Hat Virtualization Manager

The attributes are also available as Maven properties, so the downstream
packager can build the customized documentation with a command like the
following:

  $ mvn validate -Pdocument \
  -Dproduct.name='Red Hat Virtualization' \
  -Dengine.name='Red Hat Virtualization Manager'

Change-Id: Ic864ba53c57c48fd3d61b70006d00d6692515ceb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 949737773882f6cd94b6e12390258c5166366c31)