Add generic reader
Currently the metmodel doesn't provide any mechanism to read an XML object
without explicitly specifying its type. For example, to read a virtual
machine the only method available is "VmReader.readOne". But there are
situations where the SDK can't know in advance what type of object is
going to be received from the server, for example "action" or "fault".
To support that, this patch adds a new "XmlReader.read" method that decides
the type of object based on the name of the first tag, and invokes the
corresponding specific reader to read it.

Change-Id: I9949e4dcb4c127f09481d343559d12952bada940
Signed-off-by: Ondra Machacek <omachace@redhat.com>