Remote the Cluster.macPool attribute
The `macPool` attribute of the `Cluster` type was introduced when the
corresponding engine feature (see bug 994283) was targed for version 4.0
of the engine. Later this feature was re-targeted for version 4.1 of the
engine, but the specification of the API wasn't updated. This patch
updates the specification to remove the attribute.

Change-Id: I15a049c9d21f6b4c2aa1a9e0c072064ed0859901
Bug-Url: https://bugzilla.redhat.com/994283
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>