Update to metamodel 1.1.5
This version of the metamodel includes the changes needed to make the
AsciiDoc section separator configurable. By default we use the slash,
but it can be changed using the "adoc.separator" property during the
build. For example, to use a dash instead of a slash:

  $ mvn validate -Pdocument -Dadoc.separator=-

Change-Id: I645d3f391fa8d48864e06f49a4e4238bf4986bb6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 42a8a5f7038c93403fcdbef91752106c6f55d21b)