Handle appendixes correctly
Currently there is no way in the model to distinguish documents that are
appendixes from documents that aren't. The result is that appendixes may
be added to the middle of the generated document, instead of to the end,
where appendixes use to go. To avoid that this patch adds a new
'appendix' boolean attribute to the 'Document' concept, and calculates
it from the name of the document file: if the name of the document
has a prefix to indicate is order, and this prefix starts with a letter,
then it is considered an appendix. For example, the following document
will be considered an appendix:

  A00-PrimitiveTypes.adoc

And the following document will not:

  001-Authentication.adoc

Change-Id: Ic2eb27b3233ca574378e7128f81a2884e77216a9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>