Automatically fix section identifiers
Currently the separators of words inside section identifiers of
standlone documents aren't replaced by the character specified by the
'adoc.separator' property, and as a result the generated .adoc file is
rejected by tools like 'ccutil'. To avoid that this patch changes the
tool so that the identifiers are fixed automatically.

Change-Id: Idbd60d814a86aa859a4e1e5c28910cd8f9841c9f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>