Don't generate cross references to non existing sections
Currently the tool generates cross references for sections that don't
exist. This happens because some attributes only have a summary, and
thus don't have their own section, but this isn't taking into account
when generating cross references for those attributes. To avoid that
this patch changes the tool so that it checks if the attribute has its
own section, if it doesn't then the cross reference will point to the
section of the containing type, which always exist.

Change-Id: I3282623a5358f8c3b23d8e71f9af17564b3729f2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>