Separate documentation profiles
This patch separates the "document" profile into different profiles
"document", "javadoc", "pdf" and "docbook". The "document" profile will
now generate only HTML. This way the user can explicitly select the kind
of documentation to generate. This is convenient because the generation
of the PDF documentation, for example, takes a long time, and it isn't
practical to generate it when only HTML is needed.

Change-Id: I7098aa639b1dc13af1e0bf8cd28401e73a817007
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 149c4a0adad50e3602706c8d171ebbf2390e1fee)