The LogicalUnit.discardMaxSize needs to be long
The 'discardMaxSize' attribute of the 'LogicalUnit' type needs to be of
type 'long', as the default 'int' doesn't have enough range to represent
it. In the future all integer attributes will be implemented with
'BigInteger', which will solve this problem, but currently the only way
to handle this is to introduce an exception in the generator of the XML
schema.

Change-Id: Ie1873d3e50aeea45f2fbd055b7336bf9d12cae94
Related-To: https://bugzilla.redhat.com/124106
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>