Reduce the number of files generated by Publican
Currently the HTML documentation generated by Publican contains more
than 1000 HTML files, because each section has its own file, and there
is a section for each struct attribute and each method parameter. This
makes the documentation difficult to read, as most most HTML pages are
very small. This patch uses the Publican 'chunk_section_depth'
configuration parameter to limit this to two levels, so there will be
just one HTML file per type and service. That reduces the number of
files to approx 500, and the resulting HTML has a good balance between
the number of files and the size of each file.

Change-Id: Ic4b04935cc322d83e274c1dc6bd087f7f2940adc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>