Attach the generated AsciiDoc and HTML files
This patch changes the POM file so that during the build the generated
AsciiDoc and HTML files are added to a .jar file that is then attached
as the Javadoc artifact of the project. This is more correct, because
this isn't really a Java API, but an specification, so it is more
actually more correct to include as default documentation the generated
HTML instead of the Javadoc.

Change-Id: I0ff52f17c5d7e5513dd05ed9f9c887ee2830ba1a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>