Add property to enable/disable CSS linking
Currently the HTML generation is configured to generate an external
'asciidoctor.css' file containing the styles. Thisis needed when
building in Jenkins, as otherwise the styles don't work. See the commit
message for commit a63fc4e for details. But this means that the CSS is
also a separate file in the resulting documentation .jar file. That
doesn't work will with the engine, as it expects only a single
'model.html' file. To avoid that issue, and support both environments,
this patch adds a new 'adoc.linkcss' property to make this behaviour
configurable. When running in Jenkins we add '-Dadoc.linkcss=true'.
When making release builds we use '-Dadoc.linkcss=false'.

Change-Id: I4bd3e17d996be7670743da5df1d5b07b926f1b22
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 90194a6819745a02aa1f77a9c6ee9d879053afae)