Generate HTML documentation by default
Currently the HTML documentation and the artifact containing it are only
generated and attached to the build when using the 'document' profile.
But now the build of the engine requires that artifact, and developers
used to just run 'mvn clean install' will encounter a build failure. To
avoid that problem this patch changes the build process of the model so
that the HTML documetnation and the corresponding artifact are generated
by default, without requiring any optional profile.

Change-Id: I646b45cb1b6ed56d18adc343f1f66b940565021b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>