Fix formatting of default methods
Currently the default methods are formatted as one-liners:

  @DELETE
  default Response remove() { throw new UnsupportedOperationException(); }

But the JavaBuffer class expects the braces at the end and beginning of
the lines, otherwise it doesn't correctly indent the generated code. The
result is that the generated code doesn't look nice. This patch fixes
the format of default methods:

  @DELETE
  default Response remove() {
    throw new UnsupportedOperationException();
  }

This results in nicer generated source code.

Change-Id: Ia3420cd9d04dc90276e3feae4017ba99a6c095af
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>