Add 'StorageDomainDisks' and 'AttachedStorageDomainDisks'
Currently we use the same 'Disks' service definitions fo the following
disk collections:

  /disks
  /storagedomain/{sd:id}/disks
  /datacenters/{dc:id}/storagedomains/{sd:id}/disks

This is wrong, because most operations on disk are then available in
multiple places, and it prevents addition of new operations that only
make sense for disks that aren't yet part of the system. A good example
of this is the operation to register an unregistered disk. That only
makes sense for disks that appear in /storagedomains/{sd:id}/disks. To
improve the situation this patch introduces new 'StorageDomainDisks' and
'AttachedStorageDomainDisk', so each collection will have its own
service:

  /disks
  DisksService

  /storagedomain/{sd:id}/disks
  StorageDomainDisksService

  /datacenters/{dc:id}/storagedomains/{sd:id}/disks
  AttachedStorageDomainDisksService

These services will initially preserve most of the duplicated
operations, for backwards compatibility. Those duplicated operations
will be marked as deprecated, and removed in a future version of the
API.

Change-Id: I9cc2e505b4d47370c2db326fc50043064841ed2f
Related-To: https://bugzilla.redhat.com/1366960
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>