Support comparing concepts without name
Currently the comparator for objects tht implement the 'Named' interface
doesn't work for objects that have a null name. That complicates sorting
list of objects, in particular list of types that contain list types,
because list types don't have a name. To allow that this patch changes
that comparator so that it also accepts null names.

Change-Id: Ic0d05b3cf2279c2c3e3e0f450afc5158820ce476
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>