Add list types to the model
Currently list types aren't part of the list or stream returned by the
'Model.getTypes' and 'Model.types' methods. The reason for that is that
list types weren't needed for most code generation tasks. But it is now
convenient to have them there, because it is useful to know what list
types are part of the model, and what are their element types, for
example to generate XML schema list wrapper types only for the types
that really need it. This patch changes the model analyzer so that the
list types are added to the model, like all the other kinds of types.

Change-Id: I2538b881c5f683454cd8c4ff2c4fde3c47f2c56a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>