Add methods to safely convert large integers
In the generated Java source code all integers are represented using the
'BigInteger' Java class. This means that when a conversion is needed to
some other Java integer type, like 'Integer' or 'Long', the caller has
to do something like this:

  Vm vm = ...;
  long memory = vm.memory().longValue();

The problem with that is that the stored value may be larger than what
the Java 'long' type can represent, and the 'longValue()' method
performs the conversion truncating the value as needed, without warning.
The 'BigInteger' class provides methods like 'longValueExact()' which
throw an arithmetic exceptions in these cases. But the generated
exception message is too generic. To improve that this patch modifies
the code generator so that for each integer attribute it will generate a
collection of methods like these:

  Byte memoryAsByte()
  Short memoryAsShort()
  Integer memoryAsInteger()
  Long memoryAsInteger()

Those methods will perform the conversion internally, and throw an
arithmetic exception with a clearer message. For example, if the code
tries to convert the value '1000' to a byte like this:

  byte memory = vm.memoryAsByte();

The exception message will be like this:

  The integer value 1000 of the 'memory' member of the 'Vm' type
  can't be converted to a 8 bits integer because that would loss
  precision.

Change-Id: I2d58e6603da3e6deb3c0d9edd4ac7f189a340d13
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>