Skip list types in documentation report
Currently the documentation report generator creates a row for each
type. But recently the list types have also been added to the model, and
it doesn't make sense to generate rows for them, as they don't have
their own documentation. This patch changes the report generator so that
it ignores the list types.

Change-Id: I17f5a106c5c9930d0d89813c1c1f1fe64dd1861f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>