Use icons for admonitions in XML and JSON
A rencet patch modified the Asciidoctor configuration in order to render
admonitions with icons. That patch only applied to the HTML generated
directly by Asciidoctor, but not to the HTML that is included into the
XLM and JSON representations of the model. This patch addresses that
issue.

Change-Id: I7a0461cb56c4e66fd9a9139d905f060340440de2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 33e377f7ac5a8798251139e41e8f0c9ea77d409a)