Live Documentation - 'Helpers' generation
This patch introduces generation of 'Helper' classes -
one for each 'Resource' class (e.g: VmsResourceHelper).
The helper classes perform validation and signature
detection.

Simple example:
---------------

@Service
@Area("Virtualization")
public interface VmGraphicsConsolesService {

    /**
     * Add new graphics console to the virtual machine.
     *
     * @author Ondra Machacek <omachace@redhat.com>
     * @date 31 Oct 2016
     * @status added
     */
    interface Add {
        @InputDetail
        default void inputDetail() {
            mandatory(console().protocol());
        }
        @In @Out GraphicsConsole console();
    }
    ...

This helper will be generated:
------------------------------

public class VmGraphicsConsolesHelper {
    public static void validateAdd(Console console) {
        if (console==null
         || console.getProtocol()==null
        throw new IllegalArgumentException("console.getProtocol() is mandatory but was not provided.");
    }
    ...
}

There are known caevets and gaps, which will be worked out in subsequent patches:
1) Invalid input results in IlegalArgumentException (later caught and wrapped)
instead of using localization.
2) There are a handful of cases where DELETE requests have parameters. These
are not currently handled.
3) Each method 'signature' in the model is expected to have at least 1 mandatory
attribute. Right now if this is not the case - the input is ignored. In the future
this should break the build.
4) Signatures in the model are represented as extensions of the main method, e.g:
interface Approve;
interface UsingRootPassword extends Approve;
Right now the extending interface MUST ALSO BE NESTED or unexpected results will
occur. In the future this limitation should not exist.

Change-Id: Idc6d4ba2b201a9edb5012ed1806672b26ca94fad
Signed-off-by: Ori Liel <oliel@redhat.com>