Insert Live Documentation.
This patches inserts most of the API method signature
meta-data into model classes as 'Live Documentation' -
a combination of java commands and annotations that
express combinations of parameters for API operations.

The existing textual meta-data file - rsdl_metadata.yaml,
will be completely discarded down the road but for now
is still kept in the source code for referenece.

Change-Id: I9958c02e314fe04209b791ba59654871f7384471
Signed-off-by: Ori Liel <oliel@redhat.com>