Added sealing option for Template creation.
Sealing an existing template has been considered an incorrect way of
implementation. The current implementation of sealing performs it just
after the template was created and before it may be used. The sealing is
triggered by user selecting 'Seal Template' option in 'New Template'
dialog.

This patch brings REST API in line with the above behaviour. Seal
parameter to Template creation was added and Seal operation for Template
was removed, because it has never been implemented.

Change-Id: I342f6ce704a4418485e72ecd9ce2fa9109f51730
Related-To: https://bugzilla.redhat.com/1335642
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>