Explain order of 'list' isn't guaranteed
In general the order of the results returned by the 'list' methods isn't
guaranteed, except for the services that support the 'search' parameter
and for the events service. But this isn't explicitly explained in the
documentation. This patch addresses that, adding to all the 'list'
methods a sentence explaining it.

Change-Id: I05296d6730169e101fba186599c53a43c6f115b0
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 53c6b846808e48b1ebe28b14635ce93117939b56)