Clarify how to add a VM to an affinity group
Currently the example that explains how to add a virtual machine to an
affinity group uses '000-000' as the identifier for both the cluster and
the virtual machine, which can cause confusion. To avoid that this patch
changes the example to use different identifiers for all the involved
objects.

Change-Id: Id6c17bff65777aca33c5aeef0bbbb7edd06efce7
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 835450c105ac28a76b42784f1844a60f7ab44b4c)