Ignore nested methods when generationg AsciiDoc
A recent patch introduced the concept of methods that are variants, or
extensions of existing methods. This affects the calculation of the set
the generation of the documentatoin, because it results in duplicated
sectoin names, which are then rejected by the Dockbook tool. To avoid
that issue this patch changes the tool so that it ignores those methods.
This will be addressed better in the future.

Change-Id: I537dd55265a65fb2eab0f3c097df108f0d017792
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>