Introduce Event.index
Currently the Event.id attribute is used as the unique identifier of
events, and it is also used as a sequence number to implement an
implicit ordering of events. That is inconsistent with the fact that
identifiers are (usually) opaque strings. To improve that this patch
introduces a new Event.index attribute, of Integer type, that is
explicitly documented as ordered.

Change-Id: I7ac608ca14383eabfa6288afefa22da1e2b988ba
Bug-Url: https://bugzilla.redhat.com/1448511
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>