Remove underscore prefix/suffix from expressions
The underscore character is used in the model language to allow the use
of names that clash with Java reserved words. For example, if the a
parameter of a method needs to be named 'volatile', which is a Java
reserved workd, it can be defined as follows:

  @In Boolean _volatile();

Or:

  @In Boolean volatile_();

The model analyzer removes these suffix or prefix, so that the name of
the concept in the in-memory representation of the model will be
'volatile'.

This is working correctly for types, services, attributes and methods,
but it isn't working correctly for expressions. For example, the
following expressions inside an input detail specification won't be
handled correctly:

  optional(_volatile());

To fix that this patch changes the expression analyzer so that it parses
Java names using the same code than the rest of the model analyzer.

Change-Id: Ibc53bb46e236bf1011d465f0a97020f08a4571c4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>