Improve parameter validation messages
Currently the code generated for validation of parameters produces error
messages like this one:

  action.getVm().getName() is mandatory but was not provided.

Those messages will eventually be presented to the user of the API,
which doesn't need to be familiar with the Java syntax. To improve the
messages this patch changes the code generator so that they will use a
simpler syntax:

  Parameter 'vm.name' is mandatory but was not provided.

The names used in this syntax will be the sames that need to be used in
the XML and JSON representations of the objects.

Change-Id: Id4c951d2ae227e215c02341aeebef8ea9a55ebd5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>